% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormatrix_excel2.R
\name{cormatrix_excel2}
\alias{cormatrix_excel2}
\title{Easy export of correlation matrix to Excel}
\usage{
cormatrix_excel2(data, filename = "cormatrix", overwrite = TRUE, ...)
}
\arguments{
\item{data}{The data frame}

\item{filename}{Desired filename (path can be added before hand
but no need to specify extension).}

\item{overwrite}{Whether to allow overwriting previous file.}

\item{...}{Parameters to be passed to the \code{correlation} package
(see \code{?correlation::correlation})}
}
\description{
Easily output a correlation matrix and export it to
Microsoft Excel, with the first row and column frozen, and
correlation coefficients colour-coded based on effect size
(0.0-0.2: small (no colour); 0.2-0.4: medium (pink); 0.4-1.0:
large (red)), following Cohen's suggestions for small (.10),
medium (.30), and large (.50) correlation effect sizes.

Based on the \code{correlation} and \code{openxlsx} packages.
}
\examples{
\dontrun{
# Basic example
cormatrix_excel2(mtcars)
cormatrix_excel2(iris, p_adjust = "none")
cormatrix_excel2(airquality, method = "spearman")
}

}
\author{
Original author: @JanMarvin
}
\keyword{Excel}
\keyword{correlation,}
\keyword{matrix,}
