% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_mad.R
\name{find_mad}
\alias{find_mad}
\title{Identify outliers based on 3 MAD}
\usage{
find_mad(data, col.list, ID = NULL, criteria = 3)
}
\arguments{
\item{data}{The data frame.}

\item{col.list}{List of variables to check for outliers.}

\item{ID}{ID variable if you would like the outliers to be identified as such.}

\item{criteria}{How many MAD to use as threshold (similar to standard deviations)}
}
\description{
Identify outliers based on 3 median absolute deviations.

See: Leys, C., Ley, C., Klein, O., Bernard, P., & Licata, L. (2013). Detecting outliers: Do not use standard deviation around the mean, use absolute deviation around the median. Journal of Experimental Social Psychology, 49(4), 764–766. https://doi.org/10.1016/j.jesp.2013.03.013
}
\examples{
find_mad(data = mtcars,
         col.list = names(mtcars),
         criteria = 3)

mtcars2 <- mtcars
mtcars2$car <- row.names(mtcars)
find_mad(data = mtcars2,
         col.list = names(mtcars),
         ID = "car",
         criteria = 3)
}
\author{
Hugues Leduc, Charles-Étienne Lavoie, Rémi Thériault
}
\keyword{MAD,}
\keyword{mean,}
\keyword{median,}
\keyword{normalization,}
\keyword{outliers}
\keyword{standardization,}
