% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_value.R
\name{format_value}
\alias{format_value}
\alias{format_p}
\alias{format_r}
\title{Easily format p or r values}
\usage{
format_value(value, type, ...)

format_p(p, precision = 0.001)

format_r(r, precision = 0.01)
}
\arguments{
\item{type}{Value to be formatted, when using the generic \code{format_value()}}

\item{...}{To specify precision level, if necessary, when using the generic \code{format_value()}}

\item{p}{p-value to format}

\item{precision}{Level of precision desired, if necessary}

\item{r}{r-value to format}
}
\description{
Easily format p or r values. Note: converts to character class for use in figures or manuscripts to accommodate e.g., "< .001".
}
\examples{
format_value(0.00041231, "p")
format_value(0.00041231, "r")
format_p(0.0041231)
format_p(0.00041231)
format_r(0.41231)
format_r(0.041231)
}
\keyword{correlation}
\keyword{formatting,}
\keyword{p-value,}
\keyword{r-value,}
