% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_dep_version}
\alias{get_dep_version}
\title{Get required version of specified package dependency}
\usage{
get_dep_version(dep, pkg = "rempsyc")
}
\arguments{
\item{dep}{Dependency of the specified package to check}

\item{pkg}{Package to check the dependency from}
}
\description{
Get required version of specified package dependency
}
