% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{install_if_not_installed}
\alias{install_if_not_installed}
\title{Install package if not already installed}
\usage{
install_if_not_installed(pkgs)
}
\arguments{
\item{pkgs}{Packages to install if not already installed}
}
\description{
Install package if not already installed
}
