% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_assumptions.R
\name{nice_assumptions}
\alias{nice_assumptions}
\title{Easy assumptions checks}
\usage{
nice_assumptions(model, interpretation = TRUE)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Test linear regression assumptions easily with a nice summary table.
}
\examples{
# Create a regression model (using data available in R by default)
model <- lm(mpg ~ wt * cyl + gear, data = mtcars)
nice_assumptions(model)

# Multiple dependent variables at once
DV <- names(mtcars[-1])
formulas <- paste(DV, "~ mpg")
models.list <- sapply(X = formulas, FUN = lm, data = mtcars,
                      simplify = FALSE, USE.NAMES = TRUE)
assumptions.table <- do.call("rbind", lapply(models.list, nice_assumptions,
                                             interpretation = FALSE))
assumptions.table
}
\keyword{assumptions,}
\keyword{linear}
\keyword{regression,}
\keyword{statistical,}
\keyword{violations}
