% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_density.R
\name{nice_density}
\alias{nice_density}
\title{Make nice density plots easily}
\usage{
nice_density(
  variable,
  group,
  data,
  colours,
  ytitle = "Density",
  xtitle = variable,
  groups.labels = NULL,
  grid = TRUE,
  shapiro = FALSE,
  title = variable
)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Make nice density plots easily.
}
\examples{
# Make the basic plot
nice_density(variable = "Sepal.Length",
            group = "Species",
            data = iris)

# Further customization
nice_density(variable = "Sepal.Length",
            group = "Species",
            data = iris,
            colours = c("#00BA38", "#619CFF", "#F8766D"),
            xtitle = "Sepal Length",
            ytitle = "Density (vs. Normal Distribution)",
            groups.labels = c("(a) Setosa", "(b) Versicolor", "(c) Virginica"),
            grid = FALSE,
            shapiro = TRUE,
            title = "Density (Sepal Length)")
}
\keyword{density,}
\keyword{normality}
