% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_QQ.R
\name{nice_QQ}
\alias{nice_QQ}
\title{Easy per-group QQ plots}
\usage{
nice_QQ(
  variable,
  group,
  data,
  colours,
  groups.labels = NULL,
  grid = TRUE,
  shapiro = FALSE,
  title = variable
)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Easily make nice per-group QQ plots through a wrapper around the \code{ggplot2} and \code{qqplotr} packages.
}
\examples{
# Make the basic plot
nice_QQ(variable = "Sepal.Length",
       group = "Species",
       data = iris)

Further customization
nice_QQ(variable = "Sepal.Length",
       group = "Species",
       data = iris,
       colours = c("#00BA38", "#619CFF", "#F8766D"),
       groups.labels = c("(a) Setosa", "(b) Versicolor", "(c) Virginica"),
       grid = FALSE,
       shapiro = TRUE,
       title = NULL)
}
\keyword{QQ}
\keyword{distribution}
\keyword{normality,}
\keyword{plots,}
