% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_randomize.R
\name{nice_randomize}
\alias{nice_randomize}
\title{Easily randomization}
\usage{
nice_randomize(
  design = "between",
  Ncondition = 3,
  n = 9,
  condition.names = c("a", "b", "c"),
  col.names = c("id", "Condition")
)
}
\arguments{
\item{design}{The design}

\item{Ncondition}{The Ncondition}

\item{n}{The n}

\item{condition.names}{The condition.names}

\item{col.names}{The col.names}
}
\description{
Randomize easily with different designs.
}
\examples{
# Specify design, number of conditions, number of participants, and names of conditions:
nice_randomize(design = "between", Ncondition = 4, n = 8,
               condition.names = c("BP","CX","PZ","ZL"))

# Within-Group Design
nice_randomize(design = "within", Ncondition = 4, n = 6,
               condition.names = c("SV","AV","ST","AT"))

# Make a quick runsheet
randomized <- nice_randomize(design = "within", Ncondition = 4, n = 128,
                             condition.names = c("SV","AV","ST","AT"),
                             col.names = c("id", "Condition", "Date/Time",
                             "SONA ID", "Age/Gd.", "Handedness",
                             "Tester", "Notes"))
head(randomized)
}
\keyword{allocation,}
\keyword{conditions,}
\keyword{design}
\keyword{experimental}
\keyword{random}
\keyword{randomization,}
