% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_reverse.R
\name{nice_reverse}
\alias{nice_reverse}
\title{Easily recode scores}
\usage{
nice_reverse(x, max, min = 1, warning = TRUE)
}
\arguments{
\item{x}{the score to reverse}

\item{max}{the maximum score on the scale}

\item{min}{the miminum score on the scale (optional unless it isn't 1)}

\item{warning}{whether to show the warning about the minimum not being 1}
}
\description{
Easily recode scores (reverse-score), typically for questionnaire answers.
}
\examples{
# Reverse score of 5 with a maximum score of 5
nice_reverse(5, 5)

# Reverse score of 1 with a maximum score of 5
nice_reverse(1, 5)

# Reverse score of 3 with a maximum score of 5
nice_reverse(3, 5)

# Reverse score of 4 with a maximum score of 4 and a minimum of 0
nice_reverse(4, 4, min = 0)

# Reverse score of 0 with a maximum score of 4 and a minimum of 0
nice_reverse(0, 4, min = 0)

# Reverse score of -3 with a maximum score of 3 and a minimum of -3
nice_reverse(-3, 3, min = -3)

# Reverse score of 0 with a maximum score of 4 and a minimum of 0
nice_reverse(3, 3, min = -3)
}
\keyword{reverse}
\keyword{scoring}
