% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_scatter_times.R
\name{nice_scatter_times}
\alias{nice_scatter_times}
\title{Easy scatter plots over multiple times (T1, T2, T3)}
\usage{
nice_scatter_times(
  data,
  response,
  group,
  error_bars = FALSE,
  ytitle = gsub(".*_", "", response[[1]])
)
}
\arguments{
\item{data}{The data frame.}

\item{response}{The dependent variable to be plotted.}

\item{group}{The group by which to plot the variable}

\item{error_bars}{Logical, whether to include 95\% confidence intervals for means.}

\item{ytitle}{An optional x-axis label, if desired.}
}
\value{
A scatter plot of class ggplot.
}
\description{
Make nice scatter plots over multiple times (T1, T2, T3) easily.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Make the basic plot
nice_scatter_times(
  data = iris,
  response = names(iris)[1:4],
  group = "Species"
)
\dontshow{\}) # examplesIf}
}
