% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_slopes.R
\name{nice_slopes}
\alias{nice_slopes}
\title{Easy simple slopes}
\usage{
nice_slopes(
  response,
  predictor,
  moderator,
  moderator2 = NULL,
  covariates = NULL,
  data,
  ...
)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Easily compute simple slopes in moderation analysis, with effect sizes, and format in publication-ready format.
}
\examples{
# Make the basic table
nice_slopes(response = "mpg",
            predictor = "gear",
            moderator = "wt",
            data = mtcars)

# Multiple dependent variables at once
nice_slopes(response = c("mpg", "disp", "hp"),
            predictor = "gear",
            moderator = "wt",
            data = mtcars)

# Add covariates
nice_slopes(response = "mpg",
            predictor = "gear",
            moderator = "wt",
            covariates = c("am", "vs"),
            data = mtcars)

# Three-way interaction (continuous moderator and binary second moderator required)
nice_slopes(response = "mpg",
            predictor = "gear",
            moderator = "wt",
            moderator2 = "am",
            data = mtcars)
}
\keyword{interaction,}
\keyword{moderation,}
\keyword{regression}
\keyword{simple}
\keyword{slopes,}
