% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_t_test.R
\name{nice_t_test}
\alias{nice_t_test}
\title{Easy t-tests}
\usage{
nice_t_test(response, group, data, ...)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Easily compute t-test analyses, with effect sizes, and format in publication-ready format.
}
\examples{
# Make the basic table
nice_t_test(response = "mpg",
            group = "am",
            data = mtcars)

# Multiple dependent variables at once
nice_t_test(response = c("mpg", "disp", "drat", "wt", "qsec", "gear", "carb"),
            group = "am",
            data = mtcars)

# Can be passed some of the regular arguments of base `t.test()`
nice_t_test(response = "mpg",
            group = "am",
            data = mtcars,
            alternative = "less") # to make it one-sided instead of two-sided
}
\keyword{differences}
\keyword{group}
\keyword{t-test,}
