% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_table.R
\name{nice_table}
\alias{nice_table}
\title{Make nice APA tables easily}
\usage{
nice_table(
  dataframe,
  italics = NULL,
  highlight = FALSE,
  col.format.p = NULL,
  col.format.r,
  format.custom,
  col.format.custom
)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Make nice APA tables easily through a wrapper around the \code{flextable} package with sensical defaults and automatic formatting features.
}
\examples{
# Make the basic table
my_table <- nice_table(head(mtcars))
my_table

# Save table to word
save_as_docx(my_table, path = "nicetablehere.docx")

# Publication-ready tables
mtcars.std <- lapply(mtcars, scale)
model <- lm(mpg ~ cyl + wt * hp, mtcars.std)
stats.table <- as.data.frame(summary(model)$coefficients)
CI <- confint(model)
stats.table <- cbind(row.names(stats.table),
                     stats.table, CI)
names(stats.table) <- c("Term", "B", "SE", "t", "p",
                        "CI_lower", "CI_upper")
nice_table(stats.table, highlight = TRUE)

# Test different column names
test <- head(mtcars)
names(test) <- c("dR", "df", "M", "SD", "b", "np2",
                 "ges", "p", "r", "R2", "sr2")
test[, 10:11] <- test[, 10:11]/10
nice_table(test)

# Custom cell formatting (such as p or r)
nice_table(test[8:11], col.format.p = 2:4, highlight = .001)

nice_table(test[8:11], col.format.r = 1:4)

# Apply custom functions to cells
fun <- function(x) {x+11.1}
nice_table(test[8:11], col.format.custom = 2:4, format.custom = "fun")

# For the example below, I have to use a unicode
# character code instead of the multiplication symbol
# because of package documentation restrictions but you
# can use special characters directly.
fun <- function(x) {paste("\u00d7", x)}
nice_table(test[8:11], col.format.custom = 2:4, format.custom = "fun")
}
\keyword{APA}
\keyword{style}
\keyword{table}
