% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_var.R
\name{nice_var}
\alias{nice_var}
\title{Obtain variance per group}
\usage{
nice_var(variable, group, data)
}
\arguments{
\item{variable}{The variable}

\item{group}{The group}

\item{data}{The data}
}
\description{
Obtain variance per group as well as check for the rule of thumb of one group having variance four times bigger than any of the other groups.
}
\examples{
# Make the basic table
nice_var(variable="Sepal.Length",
        group="Species",
        data=iris)

# Try on multiple variables
DV <- names(iris[1:4])
var.table <- do.call("rbind", lapply(DV, nice_var, data=iris, group="Species"))
var.table

}
\keyword{variance}
