% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_varplot.R
\name{nice_varplot}
\alias{nice_varplot}
\title{Attempt to visualize variance per group}
\usage{
nice_varplot(
  variable,
  group,
  data,
  colours,
  groups.labels,
  grid = TRUE,
  shapiro = FALSE,
  ytitle = variable
)
}
\arguments{
\item{variable}{The dependent variable to be plotted.}

\item{group}{The group by which to plot the variable.}

\item{data}{The data frame}

\item{colours}{Supports providing custom colours}

\item{groups.labels}{How to label the groups?}

\item{grid}{Logical, whether to keep the default background grid or not. APA style suggests not using a grid in the background, though in this case some may find it useful to more easily estimate the slopes of the different groups.}

\item{shapiro}{Logical, whether to include the p-value from the Shapiro-Wilk test on the plot.}

\item{ytitle}{An optional y-axis label, if desired.}
}
\description{
Attempt to visualize variance per group. It is not a good way to visualize variance because the concentration of points may be hidden on such plots. Use at your own risk and for exploratory purposes only.
}
\examples{
# Make the basic plot
nice_varplot(variable = "Sepal.Length",
             group = "Species",
             data = iris)

# Further customization
nice_varplot(variable = "Sepal.Length",
             group = "Species",
             data = iris,
             colours = c("#00BA38", "#619CFF", "#F8766D"),
             ytitle = "Sepal Length",
             groups.labels = c("(a) Setosa", "(b) Versicolor", "(c) Virginica"))

}
\seealso{
Other functions useful in assumption testing: \code{\link{nice_assumptions}}, \code{\link{nice_density}}, \code{\link{nice_qq}}, \code{\link{nice_var}}
}
\keyword{variance}
