% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_violin.R
\name{nice_violin}
\alias{nice_violin}
\title{Easy violin plots}
\usage{
nice_violin(
  data,
  group,
  response,
  boot = TRUE,
  bootstraps = 2000,
  colours,
  xlabels = NULL,
  ytitle = ggplot2::waiver(),
  xtitle = NULL,
  has.ylabels = TRUE,
  has.xlabels = TRUE,
  comp1 = 1,
  comp2 = 2,
  signif_annotation = NULL,
  signif_yposition = NULL,
  signif_xmin = NULL,
  signif_xmax = NULL,
  ymin,
  ymax,
  yby = 1,
  CIcap.width = 0.1,
  obs = FALSE,
  alpha = 0.7,
  border.colour = "white"
)
}
\arguments{
\item{data}{The data}

\item{group}{The group}

\item{response}{The response}

\item{boot}{The boot}

\item{bootstraps}{The bootstraps}

\item{colours}{The colours}

\item{xlabels}{The xlabels}

\item{ytitle}{The ytitle}

\item{xtitle}{The xtitle}

\item{has.ylabels}{The has.ylabels}

\item{has.xlabels}{The has.xlabels}

\item{comp1}{The comp1}

\item{comp2}{The comp2}

\item{signif_annotation}{The signif_annotation}

\item{signif_yposition}{The signif_yposition}

\item{signif_xmin}{The signif_xmin}

\item{signif_xmax}{The signif_xmax}

\item{ymin}{The ymin}

\item{ymax}{The ymax}

\item{yby}{The yby}

\item{CIcap.width}{The CIcap.width}

\item{obs}{The obs}

\item{alpha}{The alpha}

\item{border.colour}{The border.colour}
}
\description{
Make nice violin plots easily with 95\% bootstrapped confidence intervals.
}
\examples{
# Make the basic plot
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len")

\dontrun{
# Save a high-resolution image file to specified directory
ggplot2::ggsave('niceviolinplothere.tiff', width = 7, height = 7, unit = 'in',
       dpi = 300, path = "/") # change for your own desired path
}

# Change x- and y- axes labels
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           ytitle = "Length of Tooth",
           xtitle = "Vitamin C Dosage")

# See difference between two groups
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           comp1 = "0.5",
           comp2 = "2")

nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           comp1 = 2,
           comp2 = 3)

# Compare all three groups
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           signif_annotation = c("*","**","***"), # manually enter the number of stars
           signif_yposition = c(30,35,40), # What height (y) should the stars appear?
           signif_xmin = c(1,2,1), # Where should the left-sided brackets start (x)?
           signif_xmax = c(2,3,3)) # Where should the right-sided brackets end (x)?

# Set the colours manually
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           colours = c("darkseagreen","cadetblue","darkslateblue"))

# Changing the names of the x-axis labels
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           xlabels = c("Low", "Medium", "High"))

# Removing the x-axis or y-axis titles
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           ytitle = NULL,
           xtitle = NULL)

# Removing the x-axis or y-axis labels (for whatever purpose)
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           has.ylabels = FALSE,
           has.xlabels = FALSE)

# Set y-scale manually
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           ymin = 5,
           ymax = 35,
           yby = 5)

# Plotting individual observations
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           obs = TRUE)

# Micro-customizations
nice_violin(data = ToothGrowth,
           group = "dose",
           response = "len",
           CIcap.width = 0,
           alpha = 1,
           border.colour = "black")

}
\keyword{plots}
\keyword{violin}
