% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap_circle.R
\name{overlap_circle}
\alias{overlap_circle}
\title{Interpolating the Inclusion of the Other in the Self Scale}
\usage{
overlap_circle(response, categories = c("Self", "Other"))
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Interpolating the Inclusion of the Other in the Self Scale easily.
}
\examples{
# Score of 1 (0\% overlap)
overlap_circle(1)

# Score of 3.5 (25\% overlap)
overlap_circle(3.5)

# Score of 6.84 (81.8\% overlap)
overlap_circle(6.84)

# Changing labels
overlap_circle(3.12, categories = c("Humans","Animals"))

# Saving to file (PDF or PNG)
plot <- overlap_circle(3.5)
ggsave(plot, file=NULL, width=7, height=7,
       unit='in', dpi=300) # change for your own desired path
}
\keyword{Venn}
\keyword{diagrams,}
\keyword{merging,}
\keyword{overlap,}
\keyword{psychology}
\keyword{self-other}
\keyword{social}
