% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means_over_time.R
\name{plot_means_over_time}
\alias{plot_means_over_time}
\title{Easy scatter plots over multiple times (T1, T2, T3)}
\usage{
plot_means_over_time(data, response, group, error_bars = FALSE, ytitle = NULL)
}
\arguments{
\item{data}{The data frame.}

\item{response}{The dependent variable to be plotted (e.g.,
\code{c("variable_T1", "variable_T2", "variable_T3")}, etc.).}

\item{group}{The group by which to plot the variable}

\item{error_bars}{Logical, whether to include 95\% confidence intervals for means.}

\item{ytitle}{An optional x-axis label, if desired. If \code{NULL}, will take the
variable name of the first variable in \code{response}, and keep only the part of
the string before an underscore or period.}
}
\value{
A scatter plot of class ggplot.
}
\description{
Make nice scatter plots over multiple times (T1, T2, T3) easily.
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Make the basic plot
plot_means_over_time(
  data = iris,
  response = names(iris)[1:4],
  group = "Species",
  error_bars = TRUE
)
\dontshow{\}) # examplesIf}
}
