% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rempsyc-package.R
\docType{package}
\name{rempsyc-package}
\alias{rempsyc}
\alias{rempsyc-package}
\title{rempsyc: Convenience functions for increased productivity}
\description{
Convenience functions to make your workflow faster and easier. Easily customizable plots (via `ggplot2`), nice APA tables exportable to Word (via `flextable`), easily run statistical tests or check assumptions, and automatize various other tasks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/RemPsyc/rempsyc}
  \item Report bugs at \url{https://github.com/RemPsyc/rempsyc/issues}
}

}
\author{
\strong{Maintainer}: Rémi Thériault \email{remi.theriault@mail.mcgill.ca} (\href{https://orcid.org/0000-0003-4315-6788}{ORCID})

}
\keyword{internal}
