% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_assumptions.R
\name{nice_assumptions}
\alias{nice_assumptions}
\title{Easy assumptions checks}
\usage{
nice_assumptions(model, interpretation = TRUE)
}
\arguments{
\item{model}{The \code{lm} object to be passed to the function.}

\item{interpretation}{Whether to display the interpretation helper or not.}
}
\description{
Test linear regression assumptions easily with a
nice summary table.
}
\examples{
# Create a regression model (using data available in R by default)
model <- lm(mpg ~ wt * cyl + gear, data = mtcars)
nice_assumptions(model)

# Multiple dependent variables at once
DV <- names(mtcars[-1])
formulas <- paste(DV, "~ mpg")
models.list <- lapply(X = formulas, FUN = lm, data = mtcars)
assumptions.table <- do.call("rbind", lapply(models.list, nice_assumptions,
  interpretation = FALSE
))
assumptions.table

}
\seealso{
Other functions useful in assumption testing:
\code{\link{nice_density}}, \code{\link{nice_normality}},
\code{\link{nice_qq}}, \code{\link{nice_varplot}},
\code{\link{nice_var}}. Tutorial:
\url{https://remi-theriault.com/blog_assumptions}
}
\keyword{assumptions,}
\keyword{linear}
\keyword{regression,}
\keyword{statistical,}
\keyword{violations}
