% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_contrasts.R
\name{nice_contrasts}
\alias{nice_contrasts}
\title{Easy planned contrasts}
\usage{
nice_contrasts(
  response,
  group,
  covariates = NULL,
  data,
  effect.type = "cohens.d",
  bootstraps = 2000,
  ...
)
}
\arguments{
\item{response}{The dependent variable.}

\item{group}{The group for the comparison.}

\item{covariates}{The desired covariates in the model.}

\item{data}{The data frame.}

\item{effect.type}{What effect size type to use. One of "cohens.d" (default),
"akp.robust.d", "unstandardized", "hedges.g", "cohens.d.sigma", or "r".}

\item{bootstraps}{The number of bootstraps to use for the confidence interval}

\item{...}{Arguments passed to \link[bootES:bootES]{bootES::bootES}.}
}
\value{
A dataframe, with the selected dependent variable(s), comparisons of
interest, degrees of freedom, t-values, p-values, robust Cohen's d
(dR), and the lower and upper 95\% confidence intervals of the
effect size (i.e., dR).
}
\description{
Easily compute planned contrast analyses (pairwise
comparisons similar to t-tests but more powerful when more than
2 groups), and format in publication-ready format. Supports only
three groups for the moment. In this particular case, the
confidence intervals are bootstraped on the Robust Cohen's d.
}
\details{
Statistical power is lower with the standard \emph{t} test
compared than it is with the planned contrast version for two
reasons: a) the sample size is smaller with the \emph{t} test,
because only the cases in the two groups are selected; and b)
in the planned contrast the error term is smaller than it is
with the standard \emph{t} test because it is based on all the cases
(\href{https://web.pdx.edu/~newsomj/uvclass/ho_planned\%20contrasts.pdf}{source}).

The effect size calculated here is Cohen's d (or its robust version,
if specified), as calculated by \link[bootES:bootES]{bootES::bootES}.
}
\examples{
# Basic example
nice_contrasts(
  data = mtcars,
  response = "mpg",
  group = "cyl",
  bootstraps = 200
)
\donttest{
nice_contrasts(
  data = mtcars,
  response = "disp",
  group = "gear"
)

# Multiple dependent variables
nice_contrasts(
  data = mtcars,
  response = c("mpg", "disp", "hp"),
  group = "cyl"
)

# Adding covariates
nice_contrasts(
  data = mtcars,
  response = "mpg",
  group = "cyl",
  covariates = c("disp", "hp")
)
}
}
\seealso{
Tutorial: \url{https://rempsyc.remi-theriault.com/articles/contrasts}
}
\keyword{comparisons}
\keyword{contrasts}
\keyword{differences}
\keyword{group}
\keyword{pairwise}
\keyword{planned}
