% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_lm.R
\name{nice_lm}
\alias{nice_lm}
\title{Nice formatting of lm models}
\usage{
nice_lm(model, b.label = "b", mod.id = TRUE, ...)
}
\arguments{
\item{model}{The model to be formatted.}

\item{b.label}{What to rename the default "b" column (e.g.,
to capital B if using standardized data for it to be converted
to the Greek beta symbol in the \code{nice_table} function).}

\item{mod.id}{Logical. Whether to display the model number,
when there is more than one model.}

\item{...}{Further arguments to be passed to the
\code{lm} function for the models.}
}
\description{
Formats output of \code{lm} model object for a
publication-ready format.

Note: this function uses the \code{modelEffectSizes} function
from the \code{lmSupport} package to get the sr2 effect sizes.
}
\examples{
# Make and format model
model <- lm(mpg ~ cyl + wt * hp, mtcars)
nice_lm(model)

# Make and format multiple models
model2 <- lm(qsec ~ disp + drat * carb, mtcars)
my.models <- list(model, model2)
nice_lm(my.models)

}
\seealso{
Checking simple slopes after testing for moderation:
\code{\link{nice_lm_slopes}}, \code{\link{nice_mod}},
\code{\link{nice_slopes}}. Tutorial:
\url{https://remi-theriault.com/blog_moderation}
}
\keyword{interaction,}
\keyword{moderation,}
\keyword{regression}
