% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_mod.R
\name{nice_mod}
\alias{nice_mod}
\title{Easy moderations}
\usage{
nice_mod(
  response,
  predictor,
  moderator,
  moderator2 = NULL,
  covariates = NULL,
  data,
  ...
)
}
\arguments{
\item{dataframe}{The dataframe}
}
\description{
Easily compute moderation analyses, with effect sizes, and format in publication-ready format.

Note: this function uses the \code{modelEffectSizes} function from the \code{lmSupport} package to get the sr2 effect sizes.
}
\examples{
# Make the basic table
nice_mod(response = "mpg",
        predictor = "gear",
        moderator = "wt",
        data = mtcars)

# Multiple dependent variables at once
nice_mod(response = c("mpg", "disp", "hp"),
        predictor = "gear",
        moderator = "wt",
        data = mtcars)

# Add covariates
nice_mod(response = "mpg",
         predictor = "gear",
         moderator = "wt",
         covariates = c("am", "vs"),
         data = mtcars)

# Three-way interaction
nice_mod(response = "mpg",
         predictor = "gear",
         moderator = "wt",
         moderator2 = "am",
         data = mtcars)
}
\keyword{interaction,}
\keyword{moderation,}
\keyword{regression}
