% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_na.R
\name{nice_na}
\alias{nice_na}
\title{Report missing values according to guidelines}
\usage{
nice_na(data, vars)
}
\arguments{
\item{data}{The data frame.}

\item{vars}{Variable (or list of variables) to check for NAs.}
}
\description{
Nicely reports NA values according to existing guidelines. This function reports both absolute and percentage values of specified column lists. Some authors recommend reporting item-level missing item per scale, as well as participant’s maximum number of missing items by scale. For example, Parent (2013) writes:

\emph{I recommend that authors (a) state their tolerance level for missing data by scale or subscale (e.g., “We calculated means for all subscales on which participants gave at least 75\% complete data”) and then (b) report the individual missingness rates by scale per data point (i.e., the number of missing values out of all data points on that scale for all participants) and the maximum by participant (e.g., “For Attachment Anxiety, a total of 4 missing data points out of 100 were observed, with no participant missing more than a single data point”).}

See: Parent, M. C. (2013). Handling item-level missing data: Simpler is just as good. The \emph{Counseling Psychologist}, \emph{41}(4), 568-600. https://doi.org/10.1177\%2F0011000012445176
}
\examples{
# Use whole data frame
nice_na(airquality)

# Use selected columns
nice_na(airquality,
          vars = list(c("Ozone", "Solar.R", "Wind"),
                      c("Temp", "Month", "Day")))

# This might make more sense with questionnaire items
set.seed(50)
df <- data.frame(scale1_Q1 = sample(c(NA, 1:6), replace = TRUE),
                 scale1_Q2 = sample(c(NA, 1:6), replace = TRUE),
                 scale1_Q3 = sample(c(NA, 1:6), replace = TRUE),
                 scale2_Q1 = sample(c(NA, 1:6), replace = TRUE),
                 scale2_Q2 = sample(c(NA, 1:6), replace = TRUE),
                 scale2_Q3 = sample(c(NA, 1:6), replace = TRUE),
                 scale3_Q1 = sample(c(NA, 1:6), replace = TRUE),
                 scale3_Q2 = sample(c(NA, 1:6), replace = TRUE),
                 scale3_Q3 = sample(c(NA, 1:6), replace = TRUE)
)
list.variables <- list(paste0("scale1_Q", 1:3),
                       paste0("scale2_Q", 1:3),
                       paste0("scale3_Q", 1:3))
nice_na(df, list.variables)


}
\keyword{NA,}
\keyword{guidelines}
\keyword{missing}
\keyword{values,}
