% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_na.R
\name{nice_na}
\alias{nice_na}
\title{Report missing values according to guidelines}
\usage{
nice_na(data, vars, scales)
}
\arguments{
\item{data}{The data frame.}

\item{vars}{Variable (or lists of variables) to check for NAs.}

\item{scales}{The scale names to check for NAs (single character string).}
}
\description{
Nicely reports NA values according to existing
guidelines. This function reports both absolute and percentage
values of specified column lists. Some authors recommend
reporting item-level missing item per scale, as well as
participant’s maximum number of missing items by scale.
For example, Parent (2013) writes:

\emph{I recommend that authors (a) state their tolerance level for
missing data by scale or subscale (e.g., “We calculated means
for all subscales on which participants gave at least 75\%
complete data”) and then (b) report the individual missingness
rates by scale per data point (i.e., the number of missing
values out of all data points on that scale for all participants)
and the maximum by participant (e.g., “For Attachment Anxiety,
a total of 4 missing data points out of 100 were observed,
with no participant missing more than a single data point”).}

See: Parent, M. C. (2013). Handling item-level missing
data: Simpler is just as good. The \emph{Counseling Psychologist},
\emph{41}(4), 568-600. https://doi.org/10.1177\%2F0011000012445176
}
\examples{
# Use whole data frame
nice_na(airquality)

# Use selected columns explicitly
nice_na(airquality,
  vars = list(
    c("Ozone", "Solar.R", "Wind"),
    c("Temp", "Month", "Day")
  )
)

# If the questionnaire items start with the same name, e.g.,
set.seed(15)
fun <- function() {
  c(sample(c(NA, 1:10), replace = TRUE), NA, NA, NA)
}
df <- data.frame(
  scale1_Q1 = fun(), scale1_Q2 = fun(), scale1_Q3 = fun(),
  scale2_Q1 = fun(), scale2_Q2 = fun(), scale2_Q3 = fun(),
  scale3_Q1 = fun(), scale3_Q2 = fun(), scale3_Q3 = fun()
)

# One can list the scale names directly:
nice_na(df, scales = c("scale1", "scale2", "scale3"))

}
\keyword{NA,}
\keyword{guidelines}
\keyword{missing}
\keyword{values,}
