% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_QQ.R
\name{nice_qq}
\alias{nice_qq}
\title{Easy QQ plots per group}
\usage{
nice_qq(
  variable,
  group,
  data,
  colours,
  groups.labels = NULL,
  grid = TRUE,
  shapiro = FALSE,
  title = variable
)
}
\arguments{
\item{variable}{The variable}

\item{group}{The group}

\item{data}{The data}

\item{colours}{The colours}

\item{groups.labels}{The groups.labels}

\item{grid}{The grid}

\item{shapiro}{The shapiro}

\item{title}{The title}
}
\description{
Easily make nice per-group QQ plots through a wrapper around the \code{ggplot2} and \code{qqplotr} packages.
}
\examples{
# Make the basic plot
nice_qq(variable = "Sepal.Length",
       group = "Species",
       data = iris)

# Further customization
nice_qq(variable = "Sepal.Length",
       group = "Species",
       data = iris,
       colours = c("#00BA38", "#619CFF", "#F8766D"),
       groups.labels = c("(a) Setosa", "(b) Versicolor", "(c) Virginica"),
       grid = FALSE,
       shapiro = TRUE,
       title = NULL)

}
\seealso{
Other functions useful in assumption testing: \code{\link{nice_assumptions}}, \code{\link{nice_density}}, \code{\link{nice_var}}, \code{\link{nice_varplot}}
}
\keyword{QQ}
\keyword{distribution}
\keyword{normality,}
\keyword{plots,}
