% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_t_test.R
\name{nice_t_test}
\alias{nice_t_test}
\title{Easy t-tests}
\usage{
nice_t_test(data, response, group, ...)
}
\arguments{
\item{data}{The data frame.}

\item{response}{The dependent variable.}

\item{group}{The group for the comparison.}

\item{...}{Further arguments to be passed to the \code{t.test} function (e.g., to change from two-tail to one-tail).}
}
\description{
Easily compute t-test analyses, with effect sizes, and format in publication-ready format.
}
\examples{
# Make the basic table
nice_t_test(data = mtcars,
            response = "mpg",
            group = "am")

# Multiple dependent variables at once
nice_t_test(data = mtcars,
            response = c("mpg", "disp", "drat", "wt", "qsec", "gear", "carb"),
            group = "am")

# Can be passed some of the regular arguments of base `t.test()`
nice_t_test(data = mtcars,
            response = "mpg",
            group = "am",
            alternative = "less") # to make it one-sided instead of two-sided
}
\keyword{differences}
\keyword{group}
\keyword{t-test,}
