% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_varplot.R
\name{nice_varplot}
\alias{nice_varplot}
\title{Attempt to visualize variance per group}
\usage{
nice_varplot(
  variable,
  group,
  data,
  colours,
  groups.labels,
  grid = TRUE,
  shapiro = FALSE,
  ytitle = variable
)
}
\arguments{
\item{variable}{The variable}

\item{group}{The group}

\item{data}{The data}

\item{colours}{The colours}

\item{groups.labels}{The groups.labels}

\item{grid}{The grid}

\item{shapiro}{The shapiro}

\item{ytitle}{The ytitle}
}
\description{
Attempt to visualize variance per group. It is not a good way to visualize variance because the concentration of points may be hidden on such plots. use for exploratory purposes only.
}
\examples{
# Make the basic plot
nice_varplot(variable = "Sepal.Length",
             group = "Species",
             data = iris)

# Further customization
nice_varplot(variable = "Sepal.Length",
             group = "Species",
             data = iris,
             colours = c("#00BA38", "#619CFF", "#F8766D"),
             ytitle = "Sepal Length",
             groups.labels = c("(a) Setosa", "(b) Versicolor", "(c) Virginica"))

}
\keyword{variance}
