% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_means_over_time.R
\name{plot_means_over_time}
\alias{plot_means_over_time}
\title{Easy scatter plots over multiple times (T1, T2, T3)}
\usage{
plot_means_over_time(
  data,
  response,
  group,
  error_bars = TRUE,
  ytitle = NULL,
  legend.title = "",
  print_table = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data}{The data frame.}

\item{response}{The dependent variable to be plotted (e.g.,
\code{c("variable_T1", "variable_T2", "variable_T3")}, etc.).}

\item{group}{The group by which to plot the variable}

\item{error_bars}{Logical, whether to include 95\% confidence intervals for means.}

\item{ytitle}{An optional x-axis label, if desired. If \code{NULL}, will take the
variable name of the first variable in \code{response}, and keep only the part of
the string before an underscore or period.}

\item{legend.title}{The desired legend title.}

\item{print_table}{Logical, whether to also print the computed table.}

\item{verbose}{Logical, whether to also print a note regarding the meaning
of the error bars.}
}
\value{
A scatter plot of class ggplot.
}
\description{
Make nice scatter plots over multiple times (T1, T2, T3) easily.
}
\details{
Error bars are calculated using the method of Morey (2008) through
\code{\link[Rmisc:summarySEwithin]{Rmisc::summarySEwithin()}}, but raw means are plotted instead of the normed
means. For more information, visit:
http://www.cookbook-r.com/Graphs/Plotting_means_and_error_bars_(ggplot2).
}
\examples{
\dontshow{if (requireNamespace("ggplot2", quietly = TRUE)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data <- mtcars
names(data)[6:3] <- paste0("T", 1:4, "_var")
plot_means_over_time(
  data = data,
  response = names(data)[6:3],
  group = "cyl"
)
\dontshow{\}) # examplesIf}
}
\references{
Morey, R. D. (2008). Confidence intervals from normalized data:
A correction to Cousineau (2005). \emph{Tutorials in Quantitative Methods for
Psychology}, \emph{4}(2), 61-64. \doi{10.20982/tqmp.04.2.p061}
}
